% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_$_ar1_estimation.R
\name{ar1_alter}
\alias{ar1_alter}
\title{Alternative ARIMA Application for AR(1) Model}
\usage{
ar1_alter(y)
}
\arguments{
\item{y}{A numeric vector representing the time series data.}
}
\value{
A list containing the estimated AR(1) parameter (phi_1), standard
  error, white noise standard deviation, t-value, p-value, and the approach
  used.
}
\description{
Fit an AR(1) model to the time series data using the `arima` function with
alternative parameters. The `arima` function call uses order = c(1, 0, 0)
indicating an AR(1) model with no differencing and no moving average
component. An alternative method “ML" (Maximum Likelihood) and alternative
optimization "CG" (Conjugate Gradient algorithm) arguments are used. The
maximum number of iterations is increased to 200 (i.e., `optim.control =
list(maxit = 200)`).
}
\examples{
set.seed(123)
y    <- arima.sim(n = 40, list(ar = 0.40))
y[8] <- NA
ar1_alter(y)

}
\seealso{
\code{\link{estimate_ar1}},
  \code{\link{plot_acf_pacf}},
  \code{\link{ar1_default}},
  \code{\link{ar1_moment}}
}
\keyword{internal}
