% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_$_ar1_estimation.R
\name{ar1_default}
\alias{ar1_default}
\title{Default ARIMA Application for AR(1) Model}
\usage{
ar1_default(y)
}
\arguments{
\item{y}{A numeric vector representing the time series data.}
}
\value{
A list containing the estimated AR(1) parameter (phi_1), standard
  error, white noise standard deviation, t-value, p-value, and the approach
  used.
}
\description{
Fit an AR(1) model to the time series data using the default settings of the
`arima` function. The `arima` function call uses order = c(1, 0, 0)
indicating an AR(1) model with no differencing and no moving average
component. Default methods "CSS-ML" (Conditional Sum of Squares followed by
Maximum Likelihood), default optimization "BFGS"
(Broyden-Fletcher-Goldfarb-Shanno algorithm), and default iterations are
used.
}
\examples{
set.seed(123)
y    <- arima.sim(n = 40, list(ar = 0.40))
y[8] <- NA
ar1_default(y)

}
\seealso{
\code{\link{estimate_ar1}},
  \code{\link{plot_acf_pacf}},
  \code{\link{ar1_alter}},
  \code{\link{ar1_moment}}
}
\keyword{internal}
