% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_ar1_estimation.R
\name{arima_def}
\alias{arima_def}
\title{Default ARIMA Application for AR(1) Model}
\usage{
arima_def(y, order = c(1, 0, 0))
}
\arguments{
\item{y}{A numeric vector representing the time series data.}

\item{order}{A numeric vector of length 3 specifying the order of the ARIMA
model.}
}
\value{
A list containing the estimated AR(1) parameter (phi_1), standard
  error, white noise standard deviation, t-value, p-value, and the approach
  used.
}
\description{
Fits an AR(1) model to the time series data using the default settings of
the `arima` function in R. The model specification is given by order = c(1,
0, 0), indicating an AR(1) model with no differencing and no moving average
component.
}
\examples{
set.seed(123)
y    <- arima.sim(n = 40, list(ar = 0.40))
y[8] <- NA
arima_def(y)

}
\seealso{
\code{\link{estimate_ar1}},
  \code{\link{plot_acf_pacf}},
  \code{\link{arima_alt}},
  \code{\link{arima_mm}}
}
\keyword{internal}
