% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_ar1_estimation.R
\name{arima_mm}
\alias{arima_mm}
\title{Method of Moments for AR(1) Parameters}
\usage{
arima_mm(y)
}
\arguments{
\item{y}{A numeric vector representing the time series data.}
}
\value{
A list containing the estimated AR(1) parameter (phi_1), standard error, white noise standard deviation, t-value, p-value, and the approach used.
}
\description{
In some data sets, the Maximum Likelihood (ML) method in the `arima` function
fails to provide a valid estimate for the standard error, suggesting issues
with convergence or model fit. When this happens, the method of moments (MM)
can be used. The MM approach avoids computing initial values and does not
rely on numerical optimization, making it less prone to convergence issues.
}
\details{
The MM approach is used for estimating the parameters of an autoregressive
model of order 1 (AR(1)). The primary focus is on estimating the
autoregressive coefficient \eqn{\phi_1}, its standard error, the standard
deviation of the white noise, the t-value, and the p-value.

First, the sample mean (\eqn{\bar{y}}) is computed excluding missing values.
Next, missing data are processed by identifying valid pairs where both
\eqn{y_t} and \eqn{y_{t-1}} are non-missing. These valid pairs are used to
compute the autocovariance at lag 0 (\eqn{\gamma_0}) and lag 1
(\eqn{\gamma_1}). Note that \eqn{\gamma_0} is the sample variance. The AR(1)
coefficient (\eqn{\phi_1}) is estimated as \eqn{\gamma_1 / \gamma_0}.
Residuals are then computed as the difference between observed values and
those predicted by the AR(1) model. The standard deviation of these residuals
represents the white noise standard deviation (\eqn{\sigma_\epsilon}).

The standard error of \eqn{\phi_1} (\eqn{SE_{\phi_1}}) is computed as
\eqn{\sigma_\epsilon / \sqrt{n \gamma_0}}, where \eqn{\sigma_\epsilon} and
\eqn{\gamma_0} are defined earlier and \eqn{n} is the number of valid pairs.
The t-value is calculated as \eqn{\phi_1 / SE_{\phi_1}}, and the p-value is
obtained from the t-distribution with \eqn{n - 4} degrees of freedom. \eqn{n
- 4} is used since the parameters \eqn{\bar{y}}, \eqn{\gamma_0},
\eqn{\gamma_1}, and \eqn{\phi_1} are computed in this method. Note that
\eqn{n - 2} is used for the ML method since the method does not explicitly
calculate the autocovariance parameters.
}
\examples{
set.seed(123)
y    <- arima.sim(n = 40, list(ar = 0.40))
y[8] <- NA
arima_mm(y)


}
\seealso{
\code{\link{estimate_ar1}},
  \code{\link{plot_acf_pacf}},
  \code{\link{arima_def}},
  \code{\link{arima_alt}}
}
\keyword{internal}
