% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_$_beta_logit.R
\name{beta_logit_tran}
\alias{beta_logit_tran}
\title{Beta Logit Transformation}
\usage{
beta_logit_tran(x, bl4, show_msgs = FALSE)
}
\arguments{
\item{x}{A numeric vector representing the data to be transformed.}

\item{bl4}{A numeric vector of length four providing the transformation
parameters: the lower (`a`) and upper (`b`) bounds of the data's domain, and
the `b1` and `b2` parameters of the beta distribution function.}

\item{show_msgs}{Logical; if `TRUE`, messages will be issue where applicable}
}
\value{
A numeric vector of the same length as `x`, containing the
transformed values.
}
\description{
Applies a beta logit transformation to a numeric vector `x`,
effectively scaling it from a bounded domain defined by parameters `a` and
`b` to an unbounded range. This transformation involves normalizing `x` to a
[0,1] range, applying a beta cumulative distribution function, and then using
a logit function to transform the result to the real line. This is
particularly useful for data that must be constrained within specific bounds
but requires transformation to an unbounded scale for statistical modeling or
 analysis.
}
\details{
The transformation process is defined as follows:
\enumerate{
  \item Normalize `x` to a [0,1] range based on the lower (`a`) and upper (`b`) bounds.
  \item Apply the beta cumulative distribution function with parameters `b1` and `b2`.
  \item Use the logit function to map the beta distribution values to an unbounded range.
}
This function is particularly suited for preparing bounded data for
regression models where normality or linearity assumptions on the original
scale are problematic.
}
\examples{
\dontrun{
x <- seq(0, 1, length.out = 101)
bl4 <- c(0, 1, 2, 2)  # beta logit transformation parameters
transformed_x <- beta_logit_tran(x, bl4, TRUE)
back_transformed_x <- beta_logit_tran_inverse(transformed_x, bl4, TRUE)
}

}
\seealso{
\code{\link[stats]{qbeta}} for details on beta distribution
  functions. \code{\link{beta_logit_tran_inverse}},
  \code{\link{transform_data}}, \code{\link{transform_data_inverse}}
}
\keyword{internal}
