% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_$_correlation.R
\name{compute_cor_matrix}
\alias{compute_cor_matrix}
\title{Compute Correlation Matrix and P-Values}
\usage{
compute_cor_matrix(data)
}
\arguments{
\item{data}{A data frame or matrix containing the numeric variables for which
the correlation matrix is to be computed.}
}
\value{
A list containing the following components:
\item{cor_matrix}{The correlation matrix computed using standard methods.}
\item{p_matrix}{A matrix of p-values corresponding to each element in the
correlation matrix.}
\item{mean_vector}{A vector of means corresponding to each column of data.}
}
\description{
This function computes the correlation matrix and the
  corresponding p-values for a given dataset using standard methods.
}
\examples{
data <- mtcars[, c("mpg", "disp", "hp", "qsec", "drat", "wt")]
compute_cor_matrix(data)

}
\seealso{
\code{\link{plot_cor_mat}}
}
