% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_mapping_tools.R
\name{create_buffer}
\alias{create_buffer}
\title{Create buffer around the a line}
\usage{
create_buffer(sf_obj = NA, col_name = NA, buffer_km = NA)
}
\arguments{
\item{sf_obj}{An `sf` object containing LINESTRING geometries for which the
buffer will be created.}

\item{col_name}{The name of the attribute in `sf_obj` which could be used to
specify or differentiate line features for future enhancements. Currently,
this parameter is included for compatibility and future development but is
not utilized in the function's operation.}

\item{buffer_km}{The distance (in kilometers) for the buffer zone around each
line feature. Specifying the distance allows for the creation of buffer
zones that are tailored to specific analysis requirements.}
}
\value{
An `sf` object representing the buffer zones around the line features
  specified in `sf_obj`. The geometries of the buffer zones are in the same
  CRS as the input `sf` object and are of POLYGON or MULTIPOLYGON type,
  depending on the geometry of the buffer created.
}
\description{
Creates a buffer zone of a specified distance around a line
  geometry contained within an `sf` object. This function is useful for
  spatial analysis tasks that require identifying areas within a certain
  range of a line feature, such as roads, rivers, or utility lines. The
  buffer distance is specified in kilometers and is converted internally to
  meters as `sf`'s `st_buffer` function requires distance in SI units.
}
\note{
The buffer distance is internally converted to meters; the function
  assumes planar projection. For geographic projections, consider the
  implications of this conversion and the accuracy of the buffer in
  geographic terms.
}
\examples{
# Assuming `lines` is an sf object containing LINESTRING geometries
buffer_zones <- create_buffer(sf_obj = lines, buffer_km = 1)
# Visualize the buffer zones
plot(buffer_zones)

}
\seealso{
\code{\link[sf]{st_buffer}}, for creating buffer zones around geometries.
}
