% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_$_initialization_and_work_with_lists_00.R
\name{cv}
\alias{cv}
\title{Clear the Viewer Pane in RStudio}
\usage{
cv()
}
\value{
Invisible `NULL`. The primary effect is the side effect of clearing
  the Viewer pane.
}
\description{
Clears the Viewer pane in RStudio by creating and opening a
  blank HTML file. This function is useful for resetting the Viewer pane to a
  clean state, especially after displaying multiple visualizations or HTML
  outputs in sequence. It is specifically tailored for use within the RStudio IDE.
}
\details{
The function generates a temporary directory, creates a blank HTML
  file within this directory, and then uses the RStudio Viewer to open this
  blank file, effectively clearing any previous content displayed in the
  Viewer pane. If the `rstudioapi` package is not available, the function will
  display a message indicating that it is intended for use in RStudio.
}
\examples{
\dontrun{
cv()
}

}
\seealso{
\code{\link[rstudioapi]{viewer}} for details on how to display local web content in the Viewer pane.
}
\keyword{internal}
