% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_$_time.R
\name{date_time_to_years_decimal}
\alias{date_time_to_years_decimal}
\title{Compute decimal date}
\usage{
date_time_to_years_decimal(date_chk, start_MM_DD = "01-01")
}
\arguments{
\item{date_chk}{A `POSIXct`, `POSIXlt`, or `Date` object representing the date-time.}
}
\value{
A numeric value representing the decimal date.
}
\description{
This function computes the decimal representation of a date-time object.
The decimal date is calculated as the year plus the day of the year divided by 366.
This function can handle both `POSIXct`/`POSIXlt` and `Date` types.
}
\examples{
# For POSIXct date-time
date_chk <- lubridate::ymd_hms("2024-07-30 11:15:30")
date_time_to_years_decimal(date_chk)

# For Date
date <- as.Date("2024-07-30")
date_time_to_years_decimal(date)

}
\seealso{
\code{\link{date_time_to_hours_decimal}},
  \code{\link{check_mm_dd_format}},
  \code{\link{leap_yday}}
}
\keyword{internal}
