% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_$_time.R
\name{decimal_date}
\alias{decimal_date}
\title{Compute decimal date}
\usage{
decimal_date(date_time)
}
\arguments{
\item{date_time}{A `POSIXct`, `POSIXlt`, or `Date` object representing the date-time.}
}
\value{
A numeric value representing the decimal date.
}
\description{
This function computes the decimal representation of a date-time object.
The decimal date is calculated as the year plus the day of the year divided by 366.
This function can handle both `POSIXct`/`POSIXlt` and `Date` types.
}
\examples{
# For POSIXct date-time
date_time <- lubridate::ymd_hms("2024-07-30 11:15:30")
decimal_date(date_time)

# For Date
date <- as.Date("2024-07-30")
decimal_date(date)

}
\seealso{
\code{\link{decimal_hours}},
  \code{\link{leap_yday}}
}
\keyword{internal}
