% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_$_time.R
\name{decimal_hours}
\alias{decimal_hours}
\title{Compute decimal time}
\usage{
decimal_hours(date_time)
}
\arguments{
\item{date_time}{A `POSIXct` or `POSIXlt` object representing the date-time.}
}
\value{
A numeric value representing the decimal time.
}
\description{
This function computes the decimal representation of the time of day.
The decimal time is calculated as the hour plus the minute divided by 60 and the second divided by 3600.
This function requires the input to be of type `POSIXct` or `POSIXlt`.
}
\examples{
date_time <- lubridate::ymd_hms("2024-07-30 11:15:30")
decimal_hours(date_time)

}
\seealso{
\code{\link{decimal_date}},
  \code{\link{leap_yday}}
}
\keyword{internal}
