% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_run_gam_models_baywide_00.R
\name{down_select_data}
\alias{down_select_data}
\title{Down Select Data Based on Job Configuration}
\usage{
down_select_data(job = NULL, model_data = NULL, show_msgs = FALSE)
}
\arguments{
\item{job}{A list specifying the job configuration, including the selected
subestuary, GAM by term, data source, and the date range for the analysis.}

\item{model_data}{A list containing loaded and pre-processed model data.}

\item{show_msgs}{Logical; if `TRUE`, displays messages during the data
selection process.}
}
\value{
A list containing refined datasets for use in the specified job,
including datasets for stations, the main data set, and segments.
}
\description{
Refines and filters the input datasets according to the
specifications of a given job configuration. This includes selecting segments
, stations, and data points that match the criteria defined in the job (e.g.,
subestuary, time range, and data source). The function ensures that only
relevant data is used in each specific modeling job, optimizing the modeling
process and enhancing the accuracy of the results.
}
\examples{
\dontrun{
job_config <- list(subestuary="CB1"
                   , gam_by_term="year"
                   , data_source="main"
                   , date_begin=as.Date("2010-01-01")
                   , date_end=as.Date("2010-12-31"))
refined_data <- down_select_data(job=job_config, model_data=loaded_data)
}

}
