% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_mapping_tools.R
\name{find_adj_polygon}
\alias{find_adj_polygon}
\title{Find adjoining polygons}
\usage{
find_adj_polygon(sf_poly = NA, prim_poly = NA, col_name = NA)
}
\arguments{
\item{sf_poly}{An `sf` object containing polygon geometries.}

\item{prim_poly}{A vector of names or identifiers for the primary polygons
within `sf_poly` to which adjoining polygons are to be found.}

\item{col_name}{The name of the column in `sf_poly` that contains the names
or identifiers of the polygons.}
}
\value{
Returns a vector containing the names or identifiers of polygons that
  adjoin the primary polygons specified in `prim_poly`.
}
\description{
Identifies and returns the names of polygons that are adjacent
  to a set of primary polygons within a given `sf` object. This function is
  useful for spatial analyses that require understanding the relationships
  and connectivity between different spatial features, such as identifying
  bordering regions or zones.
}
\details{
The function splits the input `sf` object into two groups based on
  the presence of the primary polygons specified. It then uses the
  `st_touches` function from the `sf` package to determine which polygons in
  the first group touch those in the second group. The result is a vector of
  names of the touching (adjoining) polygons. This can be particularly useful
  in ecological studies, urban planning, and geographical analyses where
  adjacency effects are of interest.
}
\examples{
# Assuming `sf_poly` is an `sf` object with a column "id" for polygon identifiers:
primary_polygons <- c("poly1", "poly2")
adjoining_polygons <- find_adj_polygon(sf_poly = sf_poly, prim_poly = primary_polygons, col_name = "id")
print(adjoining_polygons)

}
\seealso{
\code{\link[sf]{st_touches}}, for the function used to detect touching polygons.
}
