% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_$_transformations.R
\name{fisherZ}
\alias{fisherZ}
\title{Fisher Z Transformation}
\usage{
fisherZ(r)
}
\arguments{
\item{r}{A numeric vector representing correlation coefficients.}
}
\value{
A numeric vector of transformed values.
}
\description{
Applies the Fisher Z transformation to a numeric vector `r`.
}
\seealso{
\code{\link[stats]{qbeta}} for details on beta distribution
  functions.
  \code{\link{transform_data}},
  \code{\link{beta_logit_tran}},
  \code{\link{beta_logit_tran_inverse}},
  \code{\link{fisherZ_inverse}},
  \code{\link{transform_data_inverse}}
}
\keyword{internal}
