% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_$_transformations.R
\name{fisherZ_inverse}
\alias{fisherZ_inverse}
\title{Inverse Fisher Z Transformation}
\usage{
fisherZ_inverse(z)
}
\arguments{
\item{z}{A numeric vector representing Fisher Z values.}
}
\value{
A numeric vector of inverse transformed values (correlation coefficients).
}
\description{
Applies the inverse Fisher Z transformation to a numeric vector `z`.
}
\seealso{
\code{\link[stats]{qbeta}} for details on beta distribution functions.
  \code{\link{transform_data}},
  \code{\link{beta_logit_tran}},
  \code{\link{beta_logit_tran_inverse}},
  \code{\link{fisherZ}},
  \code{\link{transform_data_inverse}}
}
\keyword{internal}
