% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_$_one_liners.R
\name{fivenum_char}
\alias{fivenum_char}
\title{Compute Five-Number Summary as a Character String}
\usage{
fivenum_char(x, sig.fig = 4)
}
\arguments{
\item{x}{A numeric vector or a single-column data frame for which the five-number summary is to be computed.}

\item{sig.fig}{An integer specifying the number of significant figures to use in the summary. Default is 4.}
}
\value{
A character string containing the five-number summary, with values separated by commas.
}
\description{
This function computes the five-number summary of a numeric vector or a single-column data frame and returns it as a character string. The summary includes the minimum, first quartile, median, third quartile, and maximum.
}
\examples{
# Example usage
data <- c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10)
fivenum_char(data)

# Example usage with a single-column data frame
df <- data.frame(data)
fivenum_char(df)

}
\seealso{
\code{\link{fivenum_charp}}, \code{\link[stats]{fivenum}}, \code{\link[base]{signif}}
}
\keyword{internal}
