% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_$_one_liners.R
\name{fivenum_charp}
\alias{fivenum_charp}
\title{Compute Five-Number Summary with Observation Count}
\usage{
fivenum_charp(x, parens_inc = FALSE, sig.fig = 4)
}
\arguments{
\item{x}{A numeric vector or a single-column data frame for which the five-number summary and observation count are to be computed.}

\item{parens_inc}{A logical value indicating whether to include parentheses around the output. Default is `FALSE`.}

\item{sig.fig}{An integer specifying the number of significant figures to use in the five-number summary. Default is 4.}
}
\value{
A character string in the format "Obs.: n | 5-Num Summ.: min, Q1, median, Q3, max", with optional parentheses around the string if `parens_inc` is set to `TRUE`.
}
\description{
This function computes the five-number summary of a numeric vector or a single-column data frame and appends the count of non-`NA` observations. The result is returned as a formatted character string, with an option to include or exclude parentheses around the output.
}
\examples{
# Example usage
data <- c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10, NA)
fivenum_charp(data)
fivenum_charp(data, TRUE)

# Example usage with a single-column data frame
df <- data.frame(data)
fivenum_charp(df)

}
\seealso{
\code{\link{fivenum_char}}, \code{\link[stats]{fivenum}}, \code{\link[base]{signif}}
}
\keyword{internal}
