% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_objects.R
\name{gamANOVA}
\alias{gamANOVA}
\title{Prepare ANOVA table for GAM analysis}
\usage{
gamANOVA(gamo)
}
\arguments{
\item{lmo}{output from gam model}
}
\description{
Prepare ANOVA table for GAM analysis
}
\keyword{internal}
