% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_objects.R
\name{generate_date_time}
\alias{generate_date_time}
\title{generate date time}
\usage{
generate_date_time(
  start_date_time,
  stop_date_time,
  time_step = 1,
  time_units = "days",
  tz = "EST"
)
}
\arguments{
\item{start_date_time}{starting date and time}

\item{stop_date_time}{ending date and time}

\item{time_step}{time increment, Default = 1}

\item{time_units}{time units, Default = "days"}

\item{tz}{time zone, Default = "EST"}
}
\description{
generate date time
}
\keyword{internal}
