% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_$_predictions.R
\name{generate_date_time_seq}
\alias{generate_date_time_seq}
\title{Generate Date-Time Sequence}
\usage{
generate_date_time_seq(
  start_date_time,
  stop_date_time,
  time_step = 1,
  time_units = "days",
  tz = "EST"
)
}
\arguments{
\item{start_date_time}{POSIXct compatible string or object specifying the start of the date-time sequence.}

\item{stop_date_time}{POSIXct compatible string or object specifying the end of the date-time sequence.}

\item{time_step}{Numeric value indicating the step size between date-time values. Default is 1.}

\item{time_units}{Character string specifying the units for the time step. Options include "days", "hours", or "seconds". Default is "days".}

\item{tz}{Character string specifying the time zone. Default is "EST".}
}
\value{
A vector of date-time values in POSIXct format.
}
\description{
This function generates a sequence of date-time values based on the specified start and stop times,
time step, and time units.
}
\details{
The function ensures that both the start and stop date-times are in POSIXct format. It then generates
a sequence of date-time values according to the specified time step and units.
}
\examples{
# Generate date-time sequence every 2 days
generate_date_time("2024-01-01 00:00", "2024-01-10 00:00", time_step = 2, time_units = "days", tz = "EST")

# Generate date-time sequence every 3 hours
generate_date_time("2024-01-01 00:00", "2024-01-02 00:00", time_step = 3, time_units = "hours", tz = "EST")

}
