% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_cb4d_buildout_01_cut.R
\name{leap_yday_01}
\alias{leap_yday_01}
\title{Compute Day of Year Adjusted for Leap Year (TARGET DEPRECATION)}
\usage{
leap_yday_01(date_chk)
}
\arguments{
\item{date_chk}{A date or datetime object for which the adjusted day of the
year is calculated.}
}
\value{
An integer representing the day of the year, adjusted for leap years.
}
\description{
Calculates the day of the year (DOY) for a given date, adjusting
 as if every year is a leap year. This adjustment ensures consistency in day
numbering across years, particularly useful for analyses where maintaining a
consistent temporal scale is important, such as in time series modeling or
seasonal studies.
}
\examples{
\dontrun{
date_example <- as.Date("2023-03-01")
leap_yday(date_example)  # Outputs: 61
}

}
