% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_run_gam_models_baywide_00.R
\name{load_and_prep_input_data}
\alias{load_and_prep_input_data}
\title{Load and Prepare Input Data Files}
\usage{
load_and_prep_input_data(config_file, show_msgs = FALSE)
}
\arguments{
\item{config_file}{A string specifying the path to the JSON configuration
file, which outlines locations and specifications for input data files,
including water quality data, GIS segments, subestuary frameworks, and
station locations.}

\item{show_msgs}{Logical; if TRUE, prints messages during the data loading
and processing stages.
Defaults to FALSE.}
}
\value{
A list containing processed datasets: water quality data, GIS
segments, subestuary framework, and station locations, each ready for further
 analysis or modeling.
}
\description{
Loads and processes various datasets as specified in a JSON
configuration file. This function is designed to prepare the data for
modeling by loading from paths specified in the configuration file, selecting
 and mutating specific columns, and handling special cases such as negative
 values in `depth_b` within station data.
}
\examples{
\dontrun{
config_file_path <- "path/to/config_file_locations.json"
data_list <- load_and_prep_input_data(config_file_path, show_msgs = TRUE)
}

}
