% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_mapping.R
\name{map_points}
\alias{map_points}
\title{Add Colored Points to a Leaflet Map}
\usage{
map_points(
  P1 = NULL,
  sf_obj,
  obj_name,
  col_name = NULL,
  col_palette = NULL,
  legend_title = "Point Legend"
)
}
\arguments{
\item{P1}{The Leaflet map object to which points will be added; if NULL, a
new Leaflet map is created.}

\item{sf_obj}{An `sf` object containing POINT geometries.}

\item{obj_name}{The name of the attribute in `sf_obj` for popup labels.}

\item{col_name}{The name of the attribute in `sf_obj` used for coloring
points; optional.}

\item{col_palette}{A vector of colors for the different values of `col_name`;
if not provided, a default palette is used.}

\item{legend_title}{A character string for the legend title; defaults to
"Point Legend".}
}
\value{
A Leaflet map object with the added colored points and optionally a
  legend, ready for further use or display.
}
\description{
Adds colored points to a Leaflet map based on a given `sf`
  object and attribute for coloring.
}
\details{
This function enriches a Leaflet map by overlaying colored points
  derived from an `sf` object containing POINT geometries. Points can be
  dynamically colored based on a specified attribute within the `sf` object,
  enabling the visualization of categorical or numerical data. The function
  supports logical, factor, and numeric data types for the coloring
  attribute. If a specific coloring attribute and color palette are provided,
  these are used to determine the colors of the points. Otherwise, a default
  color or a reversed viridis color palette is applied, ensuring a visually
  distinct and informative map. Sorting is applied for logical coloring
  attributes to ensure proper layering of points.
}
\examples{
\dontrun{
# Assuming `stations` is an sf object with POINT geometries
map <- map_base()
map <- map_points(P1 = map, sf_obj = stations, obj_name = "name", col_name = "status",
                   col_palette = c("green", "red"), legend_title = "Station Status")
# Display the map
map
}

}
\seealso{
\code{\link[leaflet]{addCircleMarkers}}, for adding points to a Leaflet map.
\code{\link[scales]{viridis_pal}}, for details on the default color palette used.
}
