% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_mapping.R
\name{map_polygon}
\alias{map_polygon}
\title{Add Colored Polygons to a Leaflet Map}
\usage{
map_polygon(
  P1 = NULL,
  sf_obj,
  obj_name,
  col_name = NULL,
  col_palette = NULL,
  legend_title = "Polygon Legend"
)
}
\arguments{
\item{P1}{The Leaflet map object to which polygons will be added. If NULL, a
new Leaflet map will be created.}

\item{sf_obj}{An `sf` object containing MULTIPOLYGON geometries.}

\item{obj_name}{The name of the attribute in `sf_obj` to use for popup labels.}

\item{col_name}{The name of the attribute in `sf_obj` used for coloring
polygons. Optional.}

\item{col_palette}{A vector of colors for the different values of `col_name`.
If not provided, a default palette will be used.}

\item{legend_title}{A character string to be used for the legend title.
Defaults to "Polygon Legend".}
}
\value{
A Leaflet map object with the added polygons and potentially a
  legend, ready for further use or display.
}
\description{
Adds colored polygons to a Leaflet map based on a given `sf`
  object and attribute for coloring.
}
\details{
This function enhances a Leaflet map by adding colored polygons
  derived from an `sf` object. The color of each polygon can be dynamically
  set based on an attribute within the `sf` object. A default color will be
  applied if no specific coloring attribute is specified. Additionally, if a
  column name for coloring and a color palette are provided, the function
  uses these to color the polygons. Otherwise, a default viridis color
  palette is applied. This function is particularly useful for visualizing
  spatial data with varying attributes (e.g., regions of interest, zoning
  areas) directly on a map.
}
\examples{
\dontrun{
# Assuming `segments` is an sf object with MULTIPOLYGON geometries
map <- map_base()
map <- map_polygon(P1 = map, sf_obj = segments, obj_name = "region"
                 , col_name = "zone", col_palette = c("red", "green", "blue")
                 , legend_title = "Zones")
# Display the map
map
}

}
\seealso{
\code{\link[leaflet]{addPolygons}}, for adding polygons to a Leaflet map.
\code{\link[scales]{viridis_pal}}, for details on the default color palette used.
}
