% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_$_general_plotting_00.R
\name{plot_4_panel_distr}
\alias{plot_4_panel_distr}
\title{Create a 4-Panel Distribution Plot}
\usage{
plot_4_panel_distr(
  y,
  vlab = "",
  main = "Distribution Plots",
  na.rm = TRUE,
  plot_type = "all",
  n = 3000,
  bin_width_min = 0.2
)
}
\arguments{
\item{y}{A numeric vector of data to evaluate.}

\item{vlab}{A character string specifying the label for the plots. Used as
the y-axis label for the normal probability plot and boxplot, and the
x-axis label for the histogram and density plot.}

\item{main}{A character string specifying the main title for the combined
plot. Defaults to 'Distribution Plots'.}

\item{na.rm}{Logical; if `TRUE`, removes missing values before analysis.
Defaults to `TRUE`.}

\item{plot_type}{A character string specifying the type of plot ("all" or
"thin"). Defaults to `"all"`.}

\item{n}{An integer specifying the maximum number of points to plot for the
normal probability plot and density plot. Defaults to `3000`.}

\item{bin_width_min}{A numeric value specifying the minimum bin width for the
histogram and density plot. Defaults to `0.2`.}
}
\value{
A ggplot object representing the 4-panel distribution plot.
}
\description{
This function creates a 4-panel distribution plot displaying the normal probability plot,
boxplot, histogram, and density plot for a single variable.
}
\details{
The function creates a 4-panel distribution plot by combining a normal probability plot, boxplot,
histogram, and density plot for the input variable. It allows for thinning the data if the number of points
exceeds a specified maximum.
}
\examples{
\dontrun{
y <- stats::rnorm(6000)
plot_4_panel_distr(y)
plot_4_panel_distr(y, plot_type = "thin")
}

}
