% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_ar1_estimation.R
\name{plot_acf_pacf}
\alias{plot_acf_pacf}
\title{Plot Time Series, ACF, and PACF}
\usage{
plot_acf_pacf(y)
}
\arguments{
\item{y}{A numeric vector representing the time series data.}
}
\value{
This function does not return a value. It generates a series of plots.
}
\description{
This function plots a time series along with its ACF (Autocorrelation Function) and PACF (Partial Autocorrelation Function).
It also performs AR(1) parameter estimation and displays the estimated phi and its p-value in the time series plot.
}
\details{
The function first estimates the AR(1) parameters using the `estimate_ar1` function. It then sets up a layout for plotting:
\itemize{
  \item The first row contains the time series plot and lag-1 plot
  \item The second row contains the ACF and PACF plots
}
If the AR(1) estimation fails, the phi and p-value are set to NA and displayed accordingly.
}
\examples{
y <- c(0.5, 0.4, 0.6, 0.7, NA, 0.8, 0.5, 0.4, 0.6, 0.7, 0.8,
       0.5, 0.4, 0.6, 0.7, 0.8, 0.5, 0.4, 0.6, 0.7, 0.8)
plot_acf_pacf(y)

}
\seealso{
\code{\link{estimate_ar1}},
  \code{\link{arima_def}},
  \code{\link{arima_alt}},
  \code{\link{arima_mm}}
}
