% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_$_correlation.R
\name{plot_cor_mat}
\alias{plot_cor_mat}
\title{Matrix-Based Correlation Plot}
\usage{
plot_cor_mat(
  data,
  method = c("ggpairs", "corrplot"),
  sig.level = 0.05,
  color_scheme = "RdYlBu",
  color_n = 10,
  alpha = 0.35
)
}
\arguments{
\item{data}{The dataset to be used for correlation analysis.}

\item{method}{A string indicating the desired plotting method. Options are `"corrplot"` or `"ggpairs"`.}

\item{sig.level}{The significance level for indicating when to cross out panels. Default is 0.05.}

\item{color_scheme}{The color scheme to be used for both the lower and upper triangles in `corrplot.mixed`. Default is 'RdYlBu'.}

\item{color_n}{The number of colors to use in the color scale. Default is 10.}

\item{alpha}{Transparency level for scatter plots in `ggpairs`. Default is 0.35.}
}
\value{
A correlation matrix plot based on the specified method.
}
\description{
This function allows the user to create a correlation matrix plot using either
the `corrplot` or `ggpairs` approach. The function is flexible and allows the user to
customize various parameters for the plots.
}
\examples{

# Create a correlation plot using ggpairs
data <- mtcars[, c("mpg","disp", "hp", "qsec", "drat", "wt")]
plot_cor_mat(data, sig.level = 0.01)

# Create a correlation plot using corrplot
plot_cor_mat(data, method = "corrplot", sig.level = 0.01)

}
\seealso{
\code{\link{compute_cor_mat}}, \code{\link{plot_cor_mat_ggpairs}},
  \code{\link{plot_cor_mat_corrplot}}
}
