% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_$_correlation_plots.R
\name{plot_cor_mat_ggpairs_aux}
\alias{plot_cor_mat_ggpairs_aux}
\title{plot_cor_mat_ggpairs Helper Function}
\usage{
plot_cor_mat_ggpairs_aux(data, mapping, sig.level, ...)
}
\arguments{
\item{data}{The dataset containing the variables to be analyzed.}

\item{mapping}{Aesthetic mapping created by `aes` or `aes_`.}

\item{sig.level}{The significance level for indicating when to cross out panels. Default is 0.05.}
}
\description{
This function calculates and displays the correlation coefficient, p-value,
covariance, and its p-value. It adds an "X" on non-significant correlations.
}
\keyword{internal}
