% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_$_correlation_plots.R
\name{plot_corrplot}
\alias{plot_corrplot}
\title{Correlation Plot using corrplot}
\usage{
plot_corrplot(
  data,
  sig.level = 0.05,
  lower_col = "RdYlBu",
  upper_col = "RdYlBu",
  color_n = 10
)
}
\arguments{
\item{data}{The dataset to be used for correlation analysis.}

\item{sig.level}{The significance level for indicating when to cross out panels. Default is 0.05.}

\item{lower_col}{The color scheme for the lower triangle. Default is 'RdYlBu'.}

\item{upper_col}{The color scheme for the upper triangle. Default is 'RdYlBu'.}

\item{color_n}{The number of colors to use in the color scale. Default is 10.}
}
\description{
A helper function to create a correlation plot using `corrplot.mixed`.
}
\seealso{
\code{\link{plot_matrix_correlation}}, \code{\link{plot_ggpairs}}
}
\keyword{internal}
