% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_$_covariance_plots.R
\name{plot_cov_mat}
\alias{plot_cov_mat}
\title{Matrix-Based Covariance Plot}
\usage{
plot_cov_mat(data, sig.level = 0.05, alpha = 0.35)
}
\arguments{
\item{data}{A data frame containing the numeric variables to be plotted.}

\item{sig.level}{Significance level to determine if an "X" should be added to
non-significant correlations. Default is 0.05.}

\item{alpha}{Transparency level for the scatter plots in the lower diagonal.
Default is 0.35.}
}
\value{
A `ggpairs` plot displaying the covariance matrix, p-values, and
  bivariate relationships.
}
\description{
This function creates a `ggpairs` plot that displays the
  covariance matrix and p-values in the upper diagonal, bivariate scatter
  plots in the lower diagonal, and density plots on the diagonal.
}
\examples{
data <- mtcars[, c("mpg", "disp", "hp", "qsec", "drat", "wt")]
plot_cov_mat(data, sig.level = 0.02)

}
\seealso{
\code{\link{compute_cov_matrix}}
}
