% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_$_covariance_plots.R
\name{plot_cov_mat_aux}
\alias{plot_cov_mat_aux}
\title{plot_cov_mat Helper Function}
\usage{
plot_cov_mat_aux(data, mapping, calc_results, sig.level, ...)
}
\arguments{
\item{data}{A data frame containing the variables to be plotted.}

\item{mapping}{Aesthetic mapping created by `aes` or `aes_`.}

\item{calc_results}{A list containing the covariance matrix and p-value matrix as computed by `compute_cov_matrix`.}

\item{sig.level}{Significance level to determine if an "X" should be added to non-significant correlations.}

\item{...}{Additional arguments passed to `ggally_text`.}
}
\value{
A `ggplot` object displaying the covariance and p-value in the upper diagonal of a `ggpairs` plot.
}
\description{
This helper function displays the precomputed covariance and p-values on a `ggpairs` plot.
}
\seealso{
\code{\link{compute_cov_matrix}}, \code{\link{plot_cov_mat}}
}
\keyword{internal}
