% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_$_correlation_plots.R
\name{plot_matrix_correlation}
\alias{plot_matrix_correlation}
\title{Matrix-Based Correlation Plot Wrapper Function}
\usage{
plot_matrix_correlation(
  data,
  method = c("corrplot", "ggpairs"),
  sig.level = 0.05,
  color_scheme = "RdYlBu",
  color_n = 10,
  alpha = 0.35
)
}
\arguments{
\item{data}{The dataset to be used for correlation analysis.}

\item{method}{A string indicating the desired plotting method. Options are `"corrplot"` or `"ggpairs"`.}

\item{sig.level}{The significance level for indicating when to cross out panels. Default is 0.05.}

\item{color_scheme}{The color scheme to be used for both the lower and upper triangles in `corrplot.mixed`. Default is 'RdYlBu'.}

\item{color_n}{The number of colors to use in the color scale. Default is 10.}

\item{alpha}{Transparency level for scatter plots in `ggpairs`. Default is 0.35.}
}
\value{
A correlation matrix plot based on the specified method.
}
\description{
This function allows the user to create a correlation matrix plot using either
the `corrplot` or `ggpairs` approach. The function is flexible and allows the user to
customize various parameters for the plots.
}
\examples{
# Create a correlation plot using corrplot
plot_matrix_correlation(mtcars, method = "corrplot", sig.level = 0.01)

# Create a correlation plot using ggpairs
plot_matrix_correlation(mtcars, method = "ggpairs", sig.level = 0.01)

}
\seealso{
\code{\link{plot_corrplot}}, \code{\link{plot_ggpairs}}, \code{\link{plot_ggpairs_corr}}
}
