% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_$_general_plotting_00.R
\name{plot_obs_pred_res}
\alias{plot_obs_pred_res}
\title{Create and Combine Bivariate Plots for Diagnostics}
\usage{
plot_obs_pred_res(
  df,
  vars = c("y_obs", "y_pred", "y_resid"),
  vlab = c("Observed", "Predicted", "Residual"),
  main = NA,
  ...
)
}
\arguments{
\item{df}{A data frame containing the observed, predicted, and residual values.}

\item{vars}{A character vector specifying the column names for observed, predicted, and residual values.
Defaults to `c("y_obs", "y_pred", "y_resid")`.}

\item{vlab}{A character vector specifying the labels for the plots. Defaults to `c("Observed", "Predicted", "Residual")`.}

\item{main}{A character string specifying the main title for the combined plot. Defaults to `NA`.}

\item{...}{Additional arguments passed to the `plot_bivariate_data` function.}
}
\value{
A ggplot object representing the combined bivariate plots.
}
\description{
This function creates and combines bivariate plots for observed versus predicted values
and residuals versus observed values. It is a diagnostic tool to visualize the relationship between
observed, predicted, and residual values in a 1x2 grid of plots.
}
\details{
The function selects specified columns from a data frame, renames them for consistency,
and creates two bivariate plots using the `plot_bivariate_data` function. The combined plots are
displayed in a 1x2 grid with an optional title.
}
\examples{
\dontrun{
df <- data.frame(y_obs = stats::rnorm(6000), y_pred = stats::rnorm(6000), y_resid = stats::rnorm(6000))
plot_obs_pred_res(df)
plot_obs_pred_res(df, main = "Diagnostic Plots")
plot_obs_pred_res(df, main = "Diagnostic Plots", plot_type = "thin")
plot_obs_pred_res(df, main = "Diagnostic Plots", plot_type = "hex")
df1 <- data.frame(obs = stats::rnorm(1000), pred = stats::rnorm(1000), resid  = stats::rnorm(1000))
plot_obs_pred_res(df1, vars = c("obs", "pred", "resid"),
                  vlab = c("Obs.", "Pred.", "Res."),
                  main = "Diagnostic Plots", plot_type = "all",
                  smooth_line = TRUE)
}

}
\seealso{
\code{\link{plot_bivariate_data}}
}
