% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_$_initialization_and_work_with_lists_00.R
\name{pry}
\alias{pry}
\title{Extract and Assign List Components to the Parent Environment}
\usage{
pry(listvar, v = NA)
}
\arguments{
\item{listvar}{The target list or table from which components will be
extracted.}

\item{v}{A character vector of component names in `listvar` to be extracted
and assigned to variables in the parent environment.}
}
\value{
Invisibly returns `NULL`. The primary effect is the side effect of
  creating new variables in the parent environment corresponding to the
  extracted list components.
}
\description{
Extracts specified components from a list or table and assigns
  them to variables in the parent environment. This function facilitates the
  direct use of list or table components by creating variables in the current
  scope, simplifying access to nested data structures and enhancing code
  readability and efficiency.
}
\details{
`pry` iterates over the vector `v`, extracting each specified
  component from `listvar` and assigning it to a variable of the same name in
  the parent environment. This process allows for selective extraction of
  relevant data from complex or deeply nested lists or tables, making it
  easier to work with specific elements directly in the current script or
  analysis context. The function is particularly useful in scenarios where
  data stored in a list needs to be distributed across multiple variables for
  individual processing or analysis steps.
}
\examples{
my_list <- list(a = 1, b = 2, c = "three")
# Extract components a and c into the parent environment
pry(my_list, c("a", "c"))
# Now `a` and `c` are directly accessible
print(a)
print(c)

}
\seealso{
\code{\link{list}}, for creating and manipulating list objects.
\code{\link{list2env}}, for converting lists into environments.
}
\keyword{internal}
