% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_$_predictions.R
\name{replicate_rows_by_seq}
\alias{replicate_rows_by_seq}
\title{Replicate Rows by Sequence}
\usage{
replicate_rows_by_seq(
  data,
  start = 1,
  step = 1,
  replicate_col = "depth_b",
  output_col = "depth"
)
}
\arguments{
\item{data}{A data frame or tibble containing the data to be replicated.}

\item{start}{A numeric value specifying the starting point for the sequence. Default is 1.}

\item{step}{A numeric value specifying the increment for the sequence. Default is 1.}

\item{replicate_col}{A character string specifying the name of the column
used to determine the number of replications (sequence end). Default is
"depth_b".}

\item{output_col}{A character string specifying the name of the column where
the generated sequence will be stored. Default is "depth".}
}
\value{
A tibble containing the replicated rows, with a new column for the
  sequence values.
}
\description{
This function replicates rows of a data frame based on a sequence generated
from a specified column. The function takes a data frame and for each row,
generates a sequence of values and replicates the row accordingly.
}
\details{
For each row in the input data, the function generates a sequence
  starting from `start` to the value found in `replicate_col`, with
  increments of `step`. The row is then replicated for each value in the
  sequence, and a new column is created to store the sequence values.
}
\examples{
# Example data
data <- tibble::tibble(
  id = 1:2,
  depth_b = c(2, 3)
)

# Replicate rows based on the 'depth_b' column
replicate_rows_by_seq(data, start = 0.5, step = 0.5, replicate_col = "depth_b", output_col = "depth")

}
