% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_diagnostics.R
\name{report_diagnostics}
\alias{report_diagnostics}
\title{CB4D Diagnostics Report}
\usage{
report_diagnostics(
  fn_model_rda = NULL,
  dn_model_rda = NULL,
  format_rmd = "html",
  path_rmd = NULL,
  path_script = NULL,
  dn_results = file.path(tempdir(), "diagnostics_results"),
  fn_report_prefix = "cb4d_diagnostics~",
  quiet = FALSE
)
}
\arguments{
\item{fn_model_rda}{CB4D model output RDA files, single value or vector of
filenames.}

\item{dn_model_rda}{Directory containing all of the model RDA files.  If
files are in separate folders provide a vector of the same length as
fn_model_rda}

\item{format_rmd}{Notebook (RMD) output format, html or word.
Default = "html"}

\item{path_rmd}{Path (directory and file name) of Notebook (RMD) file.}

\item{path_script}{Path (directory and filename) of script to create objects
for use in Notebook (RMD).  If NULL then only objects in the global
environment will be used for the Notebook.  Default = NULL}

\item{dn_results}{Directory to save created notebook(s). The directory will
be created if it does not exist.  Default is tempdir()}

\item{fn_report_prefix}{Prefix for generated Notebook output.  The base name
of the file will be the base name of the input RDA file.
Default = "cb4d_diagnostics~"}

\item{quiet}{Should printing of messages be suppressed.  This includes this
function and during rendering of the document (rmarkdown::render).
Default = TRUE}
}
\description{
Create diagnostics report for CB4D model output
}
\details{
Default output directory is temporary directory
Can access in Windows from the R console via `shell.exec(tempdir())`
}
\examples{

# Example 1, Select File(s), file name
fn_model_rda     <- "gs~CB5~all~2010_2022~mean~2023_10_23_005~by_term2.rda"
dn_model_rda     <- file.path("cb4d_model", "cb4d_gam_models")

# Other input parameters
format_rmd       <- "html"
path_rmd         <- file.path("diagnotics_scripts"
                              , "cb4d_model_diagnostics_vol01_report.Rmd")
path_script      <- file.path("diagnotics_scripts"
                              , "cb4d_model_diagnostics_vol01_objects.R")
# use defaults for last 3 parameters
dn_results       <- file.path(tempdir(), "diagnostics_results")
fn_report_prefix <- "cb4d_model_report_vol01~"
quiet            <- FALSE

# Run Function
report_diagnostics(fn_model_rda
                  , dn_model_rda
                  , format_rmd
                  , path_rmd
                  , path_script
                  , dn_results
                  , fn_report_prefix
                  , quiet)
# open results folder (Windows only)
shell.exec(dn_results)


# Example 2, Select File(s), dialog box (Windows only)
if (!require(tcltk)) {install.packages("tcltk")}  #install if needed
library(tcltk)
## get full path so have to convert for use with function
path_model_rda  <- tcltk::tk_choose.files()
fn_model_rda    <- basename(path_model_rda)
dn_model_rda    <- dirname(path_model_rda)

# Other input parameters
format_rmd       <- "html"
path_rmd         <- file.path("diagnotics_scripts"
                              , "cb4d_diagnostics_vol01_report.Rmd")
path_rmd <- file.path("inst", "rmd", "cb4d_model_diagnostics_vol01_report.Rmd")
path_script      <- file.path("diagnotics_scripts"
                              , "cb4d_diagnostics_vol01_objects.R")
# use defaults for last 3 parameters
dn_results       <- file.path(tempdir(), "diagnostics_results")
fn_report_prefix <- "cb4d_model_report_vol01~"
quiet            <- FALSE

# Run Function
report_diagnostic(fn_model_rda
                  , dn_model_rda
                  , format_rmd
                  , path_rmd
                  , path_script
                  , dn_results
                  , fn_report_prefix
                  , quiet)
# open results folder (Windows only)
shell.exec(dn_results)

}
