% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_objects.R
\name{report_objects}
\alias{report_objects}
\title{CB4D Diagnostics Report, Create Objects}
\usage{
report_objects(report_version = NULL)
}
\arguments{
\item{report_version}{Version of report (1 or 2).  Default = NULL}
}
\description{
Create objects for diagnostics report for CB4D model
}
\details{
Creates objects for use with the diagnostic report.

Input is a path and file name.  The file is a script that creates the objects
used in the report.  The data it uses must be in the global environment.

This function replaces the "source" command in the original version of the
scripts to create the report objects.

Objects for volume "1" and "2" only.
}
\keyword{internal}
