% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_run_gam_models_baywide_00.R
\name{rg_model_statistics}
\alias{rg_model_statistics}
\title{Calculate Model Statistics}
\usage{
rg_model_statistics(gs, df)
}
\arguments{
\item{gs}{The fitted model object from which statistics will be calculated.}

\item{df}{A data frame containing the observed values and residuals from the
model.}
}
\value{
A tibble with calculated statistical metrics for the model, including
 RMSE, R-squared, and AIC.
}
\description{
Computes key statistical metrics for a model based on residuals
and observed values. This function calculates the root mean square error
(RMSE), R-squared value, and Akaike Information Criterion (AIC) for a given
set of model outputs, aiding in the evaluation of model performance.
}
\examples{
\dontrun{
stats <- rg_model_statistics(fitted_model, model_data_frame)
}

}
\keyword{internal}
