% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_run_gam_models_baywide_00.R
\name{run_gam}
\alias{run_gam}
\title{Execute GAM Modeling Job}
\usage{
run_gam(
  job = NULL,
  data = NULL,
  model_config = NULL,
  post_process = TRUE,
  show_msgs = TRUE
)
}
\arguments{
\item{job}{A list detailing the job configuration, including the model
formula, data transformations, and variables.}

\item{data}{A data frame containing the dataset to be used in the model.}

\item{model_config}{A list containing the model configuration settings.}

\item{post_process}{Logical; if `TRUE`, the function will append predictions
and calculate residuals.}

\item{show_msgs}{Logical; if `TRUE`, progress messages will be displayed.}
}
\value{
The fitted model object, potentially augmented with additional data
such as predictions and residuals if post-processing is enabled.
}
\description{
Facilitates the execution of a Generalized Additive Model (GAM)
job based on specified #' parameters and data. This function sets up the data
, applies any required transformations, runs the GAM, and optionally
post-processes the output to include predictions and residuals. It is
designed to streamline the modeling process, from data preparation through to
model evaluation.
}
\examples{
\dontrun{
fitted_model <- run_gam(job_config, data_frame, model_config_settings)
}

}
