% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_$_initialization_and_work_with_lists_00.R
\name{stow}
\alias{stow}
\title{Append Variables to a List}
\usage{
stow(listvar = list(), v = NA)
}
\arguments{
\item{listvar}{A list to which variables will be appended. If not provided,
an empty list is initialized.}

\item{v}{A character vector containing the names of variables to append to
`listvar`.}
}
\value{
A list containing the original elements of `listvar` with the
  additional variables appended.
}
\description{
Dynamically appends variables to an existing list based on the
  names provided in the input vector `v`. This function evaluates the names
  in `v` within the current environment and appends the corresponding objects
  to the list `listvar`. It's useful for aggregating multiple variables into
  a single list structure dynamically, facilitating data management and
  manipulation in complex workflows.
}
\details{
The function iterates over the vector `v`, evaluating each name to
  find the corresponding object in the current environment. These objects are
  then appended to the list `listvar` under their original names. This
  approach allows for flexible and dynamic construction of lists from
  separate variables without manually specifying each one. It is particularly
  handy in scenarios where variable names or objects are generated
  programmatically, and direct aggregation into a list simplifies subsequent
  data processing steps.
}
\examples{
# Define some variables
a <- 1; b <- 2; c <- "three"
# Initialize an empty list
my_list <- list()
# Append variables a and b to the list
my_list <- stow(my_list, c("a", "b"))
print(my_list)

}
\seealso{
\code{\link{list}}, for creating lists.
\code{\link{eval}}, \code{\link{parse}}, for details on evaluating and parsing expressions.
}
\keyword{internal}
