% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_cb4d_buildout_00.R
\name{tblFT1}
\alias{tblFT1}
\title{Output Formatted Table Using Flextable}
\usage{
tblFT1(
  data,
  tbl_title = NA,
  tbl_pre_label = NA,
  tbl_font_name = NA,
  tbl_font_size = NA,
  tbl_theme = NA
)
}
\arguments{
\item{data}{A dataframe or tibble to be displayed as a formatted table.}

\item{tbl_title}{(Optional) The title of the table to be displayed above it.}

\item{tbl_pre_label}{(Optional) Prefix label for the table title, e.g., "Table:".}

\item{tbl_font_name}{(Optional) The font name to be used throughout the table.}

\item{tbl_font_size}{(Optional) The font size to be applied to the table text.}

\item{tbl_theme}{(Optional) The theme to apply to the table. Supported themes
include "box", "vanilla", and "booktabs".}
}
\value{
A `flextable` object representing the formatted table. This object
can be printed in the R console, rendered in the RStudio Viewer, or included
in R Markdown documents.
}
\description{
Outputs a formatted table based on the provided dataset.
  Utilizes the `flextable` package to create highly customizable tables with
  options for title, font, size, and theme adjustments. This function is
  designed for both interactive use within RStudio and integration within R
  Markdown documents, offering a versatile solution for displaying data in a
  visually appealing format.
}
\details{
The function offers several customization options:
\itemize{
  \item Table title and prefix label can be specified, allowing for
  descriptive headers above the table.
  \item Font name and size customization enable adherence to specific styling
  guidelines or preferences.
  \item A selection of table themes (`"box"`, `"vanilla"`, `"booktabs"`)
  allows for varied stylistic presentations of the data.
}
Additionally, the function detects the running environment to appropriately
render the table for interactive previews or for embedding within R Markdown
documents. This adaptability makes it an effective tool for data presentation
in diverse contexts.
}
\examples{
\dontrun{
tblFT1(head(iris), tbl_title = "Summary of Iris Data", tbl_theme = "box")
}

}
\seealso{
\code{\link[flextable]{flextable}}, for more details on creating and customizing flextables.
\code{\link[officer]{run_autonum}}, for automatically numbering tables and figures in R Markdown.
}
