% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidBIT.R
\docType{package}
\name{tidBIT-package}
\alias{tidBIT}
\alias{tidBIT-package}
\title{tidBIT: Chesapeake Bay 4D Model Functions}
\description{
Collection of helper functions for Chesapeake Bay 4d model.
}
\details{
This software program is preliminary or provisional and is subject
to revision. This software program is for testing only, no warranty,
expressed or implied, is made as to the accuracy and functioning of the
program and related program material nor shall the fact of distribution
constitute any such warranty, and no responsibility is assumed in connection
therewith. This software is provided 'AS IS.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tetratech/tidBIT}
  \item Report bugs at \url{https://github.com/tetratech/tidBIT/issues}
}

}
\author{
\strong{Maintainer}: Erik W. Leppo \email{Erik.Leppo@tetratech.com}

Authors:
\itemize{
  \item Jon Harcum \email{Jon.Harcum@tetratech.com}
}

}
