% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_ar1_estimation.R
\name{ar1_alter}
\alias{ar1_alter}
\title{Alternative ARIMA Application for AR(1) Model}
\usage{
ar1_alter(y)
}
\arguments{
\item{y}{A numeric vector representing the time series data.}
}
\value{
A list containing the estimated AR(1) parameter (phi_1), standard
  error, white noise standard deviation, t-value, p-value, and the approach
  used.
}
\description{
Alternative ARIMA Application for AR(1) Model
}
\examples{
set.seed(123)
y    <- arima.sim(n = 40, list(ar = 0.40))
y[8] <- NA
ar1_alter(y)

}
\seealso{
\code{\link{estimate_ar1}},
  \code{\link{plot_acf_pacf}},
  \code{\link{ar1_default}},
  \code{\link{ar1_moment}}
}
\keyword{internal}
