% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_ar1_estimation.R
\name{arima_alt}
\alias{arima_alt}
\title{Alternative ARIMA Application for AR(1) Model}
\usage{
arima_alt(y, order = c(1, 0, 0))
}
\arguments{
\item{y}{A numeric vector representing the time series data.}

\item{order}{A numeric vector of length 3 specifying the order of the ARIMA model.}
}
\value{
A list containing the estimated AR(1) parameter (phi_1), standard
  error, white noise standard deviation, t-value, p-value, and the approach
  used.
}
\description{
Fits an AR(1) model to the time series data using modified settings of the
`arima` function in R. This includes using method "ML" and optimization
method "CG".
}
\examples{
set.seed(123)
y    <- arima.sim(n = 40, list(ar = 0.40))
y[8] <- NA
arima_alt(y)

}
\seealso{
\code{\link{estimate_ar1}},
  \code{\link{plot_acf_pacf}},
  \code{\link{arima_def}},
  \code{\link{arima_mm}}
}
\keyword{internal}
