% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_beta_logit.R
\name{beta_logit_tran_inverse}
\alias{beta_logit_tran_inverse}
\title{Inverse Beta Logit Transformation}
\usage{
beta_logit_tran_inverse(z, bl4, show_msgs = FALSE)
}
\arguments{
\item{z}{A numeric vector representing the transformed data to be
back-transformed.}

\item{bl4}{A numeric vector of length four providing the transformation
parameters: the lower (`a`) and upper (`b`) bounds of the original data's
domain, and the `b1` and `b2` parameters of the beta distribution function.}

\item{show_msgs}{Logical; if `TRUE`, messages will be issue where applicable}
}
\value{
A numeric vector of the same length as `z`, containing the
back-transformed values, aiming to reflect the original observations.
}
\description{
Reverses the beta logit transformation applied by
`beta_logit_tran`, converting a numeric vector `z` from an unbounded range
back to its original bounded domain. This process involves applying the
inverse logit function to `z`, using the inverse beta cumulative distribution
 function, and then scaling the result back to the original domain defined by
 parameters `a` and `b`. It is useful for interpreting the results of
analyses performed on data transformed by `beta_logit_tran`, allowing the
transformed values to be mapped back to their original scale.
}
\details{
The inverse transformation process consists of the following steps:
\enumerate{
  \item Apply the inverse logit function to `z`, mapping it from an unbounded range to (0,1).
  \item Use the inverse beta cumulative distribution function with parameters `b1` and `b2` to
       find the corresponding quantile values.
  \item Scale these values back to the original domain of the data, defined by the lower (`a`)
       and upper (`b`) bounds.
}
This function enables the recontextualization of model outputs or analyses
back to the practical, bounded domain of the original data, facilitating
interpretation and decision-making.
}
\examples{
\dontrun{
z <- seq(-5, 5, length.out = 100)
bl4 <- c(0, 1, 2, 2)  # Lower and upper bounds with beta distribution parameters
original_x <- beta_logit_tran_inverse(z, bl4)
}

}
\seealso{
\code{\link[stats]{qbeta}}, for details on the inverse beta distribution function.
\code{\link{expit}}, for information on the inverse logit function.
}
