% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_cb4d_buildout_00.R
\name{cv}
\alias{cv}
\title{Clear the Viewer Pane in RStudio}
\usage{
cv()
}
\value{
Invisible `NULL`. The primary effect is the side effect of clearing
  the Viewer pane.
}
\description{
Clears the Viewer pane in RStudio by creating and opening a
  blank HTML file. This function is useful for resetting the Viewer pane to a
  clean state, especially after displaying multiple visualizations or HTML
  outputs in sequence. It leverages the `rstudioapi` package to interact with
  the RStudio environment, making it specifically tailored for use within
  RStudio IDE.
}
\details{
The function generates a temporary directory, creates a blank HTML
  file within this directory, and then uses the RStudio Viewer to open this
  blank file, effectively clearing any previous content displayed in the
  Viewer pane. This is particularly handy in interactive R sessions and
  RMarkdown documents when an empty Viewer pane is desired for either
  aesthetic or functional reasons.
}
\examples{
\dontrun{
cv()
}

}
\seealso{
\code{\link[rstudioapi]{viewer}}, for details on how to display local web content in the Viewer pane.
}
