% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_$_ar1_estimation.R
\name{estimate_ar1}
\alias{estimate_ar1}
\title{Estimate AR(1) Model Parameter}
\usage{
estimate_ar1(y)
}
\arguments{
\item{y}{A numeric vector representing the time series data.}
}
\value{
A list containing the following elements:
\describe{
  \item{phi_1}{The estimated AR(1) parameter.}
  \item{standard_error}{The standard error of the estimated AR(1) parameter.}
  \item{white_noise_sd}{The standard deviation of the white noise.}
  \item{t_value}{The t-value of the estimated AR(1) parameter.}
  \item{p_value}{The p-value of the estimated AR(1) parameter.}
  \item{approach}{The method used to estimate the AR(1) parameter ("default", "alternative", "moments", or "No method worked").}
  \item{arima_results}{Results from arima function call when either the "default" or "alternative" method were.}
  \item{message}{Any warning or error messages encountered during the estimation process.}
}
}
\description{
An AR(1) model is fit using up to three different approaches sequentially. If
an approach fails (produces an error or warning), the next approach is
attempted. Approach 1 uses the `arima` function with default parameters and
is computationally faster than Approach 2 but sometimes fails to converge.
Approach 2 uses the `arima` function with alternative method, optimization,
and iteration settings and based on testing found to be
less prone to convergence issues. Approach 3 uses the method of moments but
is considered less accurate than the other two approaches but does not rely
on computing initial values. See \code{\link{ar1_default}}, \code{\link{ar1_alter}},
and \code{\link{ar1_moment}} for further details.
}
\examples{
set.seed(123)
y    <- arima.sim(n = 40, list(ar = 0.40))
y[8] <- NA
plot_acf_pacf(y)
estimate_ar1(y)


}
\seealso{
\code{\link{plot_acf_pacf}},
  \code{\link{ar1_default}},
  \code{\link{ar1_alter}},
  \code{\link{ar1_moment}}
}
