% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_ar1_estimation.R
\name{estimate_ar1}
\alias{estimate_ar1}
\title{Estimate AR(1) Model Parameter}
\usage{
estimate_ar1(y, order = c(1, 0, 0))
}
\arguments{
\item{y}{A numeric vector representing the time series data.}

\item{order}{A numeric vector of length 3 specifying the order of the ARIMA model.}
}
\value{
A list containing the following elements:
\describe{
  \item{phi_1}{The estimated AR(1) parameter.}
  \item{standard_error}{The standard error of the estimated AR(1) parameter.}
  \item{white_noise_sd}{The standard deviation of the white noise.}
  \item{t_value}{The t-value of the estimated AR(1) parameter.}
  \item{p_value}{The p-value of the estimated AR(1) parameter.}
  \item{approach}{The method used to estimate the AR(1) parameter ("default", "alternative", "moments", or "No method worked").}
  \item{message}{Any warning or error messages encountered during the estimation process.}
}
}
\description{
This function attempts to estimate the parameters of an AR(1) model using
three different methods sequentially: default ARIMA, alternative ARIMA,
and method of moments. If a method fails (produces an error or warning),
the next method is attempted. The function returns a list containing
the estimated parameters and any warning or error messages encountered.
}
\examples{
set.seed(123)
y    <- arima.sim(n = 40, list(ar = 0.40))
y[8] <- NA
plot_acf_pacf(y)
estimate_ar1(y)


}
\seealso{
\code{\link{plot_acf_pacf}},
  \code{\link{arima_def}},
  \code{\link{arima_alt}},
  \code{\link{arima_mm}}
}
