% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_$_time.R
\name{leap_yday}
\alias{leap_yday}
\title{Compute Day of Year Adjusted for Leap Year and Start Date}
\usage{
leap_yday(date_chk, start_MM_DD = "01-01")
}
\arguments{
\item{date_chk}{A date or datetime object, or a vector of dates, for which the adjusted day of the
year is calculated.}

\item{start_MM_DD}{A character string specifying the start date in "MM-DD" format.
Default is "10-01" (water year starting Oct 1).}
}
\value{
An integer vector representing the day of the year, adjusted for leap years
and the specified start date.
}
\description{
Calculates the day of the year (DOY) for a given date, adjusting
 as if every year is a leap year. This adjustment ensures consistency in day
 numbering across years, particularly useful for analyses where maintaining a
 consistent temporal scale is important, such as in time series modeling or
 seasonal studies. The function allows adjusting for different start dates.
}
\examples{
date_example <- as.Date(
  c("1999-01-01", "1999-01-02", "1999-02-28", NA,
    "1999-03-01", "1999-03-30", "1999-03-31", "1999-04-01",
    "1999-04-02", "1999-09-29", "1999-09-30", "1999-10-01",
    "1999-10-02", "1999-12-30", "1999-12-31",
    "2000-01-01", "2000-01-02", "2000-02-28", "2000-02-29",
    "2000-03-01", "2000-03-30", "2000-03-31", "2000-04-01",
    "2000-04-02", "2000-09-29", "2000-09-30", "2000-10-01",
    "2000-10-02", "2000-12-30", "2000-12-31",
    "2001-01-01", "2001-01-02", "2001-02-28", NA,
    "2001-03-01", "2001-03-30", "2001-03-31", "2001-04-01",
    "2001-04-02", "2001-09-29", "2001-09-30", "2001-10-01",
    "2001-10-02", "2001-12-30", "2001-12-31"))

x <- leap_yday(date_example)
matrix(x, ncol = 3)

x <- leap_yday(date_example, start_MM_DD = "10-01")
matrix(x, ncol = 3)

x <- leap_yday(date_example, start_MM_DD = "04-01")
matrix(x, ncol = 3)

}
\seealso{
\code{\link{decimal_date}},
  \code{\link{decimal_hours}}
}
\keyword{internal}
